+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@electric-sql/client": "^1.0.7",
+    "@tailwindcss/postcss": "^4.1.12",
     "@tailwindcss/vite": "^4.0.6",
     "@tanstack/electric-db-collection": "^0.0.15",
     "@tanstack/react-db": "^0.0.33",
@@ -33,9 +34,10 @@
     "drizzle-orm": "^0.44.3",
     "drizzle-zod": "^0.7.1",
     "pg": "^8.16.3",
+    "postcss": "^8.5.6",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "tailwindcss": "^4.0.6",
+    "tailwindcss": "^4.1.11",
     "vite": "^6.3.3",
     "vite-tsconfig-paths": "^5.1.4",
     "zod": "^3.25.76"
