+++ src/router.tsx
@@ -3,7 +3,7 @@ import { createRouter as createTanstackRouter } from "@tanstack/react-router"
 // Import the generated route tree
 import { routeTree } from "./routeTree.gen"
 
-import "./styles.css"
+import "./styles/styles.css"
 
 // Create a new router instance
 export const createRouter = () => {
