+++ src/routes/__root.tsx
@@ -6,7 +6,7 @@ import {
 } from "@tanstack/react-router"
 import { TanStackRouterDevtools } from "@tanstack/react-router-devtools"
 
-import appCss from "../styles.css?url"
+import appCss from "../styles/styles.css?url"
 
 export const Route = createRootRoute({
   head: () => ({
