+++ tsconfig.json
@@ -1,5 +1,11 @@
 {
-  "include": ["src/**/*.ts", "src/**/*.tsx", "*.ts", "*.config.ts"],
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    "*.ts",
+    "*.config.ts",
+    "src/styles/**/*.css"
+  ],
   "exclude": ["node_modules", "dist", "build", ".vite"],
   "compilerOptions": {
     "target": "ES2022",
